Separating sea-land ridges to make mountain groupings
This change attempts to continue the work looking at ridges that combine the mountains as a source to determine mountain ranges. In this case we break apart the ridges that combine two peaks where 1 peak is above water and the other peak is below. This works okay but has some limitations -- I'll continue iterating on the peaks.